Added tests for services

+++ internal/post-service/delete.go
@@ -3,6 +3,7 @@ package postservice
 import (
 	authservice "PlantSite/internal/auth-service"
 	"context"
+	"fmt"
 
 	"github.com/google/uuid"
 )
@@ -15,6 +16,8 @@ func (s *PostService) Delete(ctx context.Context, id uuid.UUID) error {
 	if !user.HasAuthorRights() {
 		return ErrNotAuthor
 	}
+	if id == uuid.Nil {
+		return fmt.Errorf("nil post")
+	}
 	return s.postRepo.Delete(ctx, id)
 }
-
